requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
diff-match-patch==20241021
Django==5.1.1
django-cors-headers==4.6.0
django-import-export==4.3.5
django-simpleui==2025.1.13
django_neomodel==0.2.0
djangorestframework==3.15.2
idna==3.10
mysqlclient==2.2.7
neo4j==5.19.0
neomodel==5.3.3
pillow==11.1.0
PyJWT==2.10.1
pytz==2025.1
requests==2.32.3
sqlparse==0.5.1
tablib==3.8.0
tzdata==2024.1
urllib3==2.3.0

python-dotenv~=1.0.1